Revert "Fix compiler warnings"
This reverts commit 647b7b191c1a5d1a43e660ea68aa35579b16b555.

+++ src/main/formats/ChunSnesSeq.cpp
@@ -74,7 +74,7 @@ bool ChunSnesSeq::GetHeaderInfo(void) {
     header->AddSimpleItem(curOffset, 2, trackName.str());
 
     ChunSnesTrack *track = new ChunSnesTrack(this, addrTrackStart);
-    track->index = static_cast<uint8_t>(aTracks.size());
+    track->index = aTracks.size();
     aTracks.push_back(track);
 
     curOffset += 2;
